[AND-404] Create generic Mintegral ad UI composable wrapper
